src/producer/requirements.txt
nvidia-ml-py==13.580.82
paho-mqtt==2.1.0
psutil==7.1.0
pydantic==2.12.0